refactor: Detail 관련 reducer 분리
